backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==21.1
drf-extra-fields==3.4.0
djoser==2.2.0
drf_yasg
django-phonenumber-field[phonenumbers]
djoser
Pillow==9.0.0
pytest
